Refactor database handling and remove Develop Strategy V2
This commit includes significant changes to the database handling strategy, shifting from using the `DatabaseConnection` class to directly using `SQLCommand` and `SQLConnection` classes for database operations. The `AgentData` class has been modified accordingly. The `SQLCommand` class in `SQLCommand - Org.cpp` and `SQLCommand` class in `Org.h` have been removed, indicating a significant change in SQL operations handling. The `SQLCommand` and `SQLConnection` classes have been significantly refactored, with a shift from OLE DB to ODBC for managing the connection to a SQL data source. Diagnostic logging code has been added in `Run.DevelopStrategy_V2.cpp`. The version 2 of the `Develop Strategy` (`Develop Strategy V2`) is being removed or refactored from the project. The `SQLConnection` class has been updated with new methods and constructors. A new class `SQLCommand` has been created in the `database` namespace to represent a Transact-SQL statement or stored procedure to execute against a SQL Server database.
