+++ SOS/Domain/Develop Strategy/Run.DevelopStrategy.h
@@ -1,20 +1,8 @@
 #pragma once
 
-#include <valarray>
-#include <functional>
-#include "..\..\Utilities\ThreadSafeArray_V2.h"
-#include "..\..\DataStore\OrderMatrix.h"
-#include "..\..\Plush\Environment.h"
-
 namespace domain
 {
 	namespace develop_strategy
 	{
-		//extern Utilities::ThreadSafeArray_2D_V2<unsigned long> orders;
-		extern datastore::OrderMatrix order_matrix;
-
-		void compute_training_orders_thread_safe(Plush::Environment& _env, std::function<std::tuple<double, unsigned long>(Plush::Environment& env, unsigned int strategy_index, unsigned long case_index)> _run_strategy_threadsafe);
-
-		int run();
 	}
 }
